+++ src/test/groovy/ru/vyarus/gradle/plugin/lib/UpstreamKitTest.groovy
@@ -274,18 +274,18 @@ rootProject.name = "test"
         test.exists()
 
         when: "run coverage task"
-        result = runVer(GRADLE_VERSION, 'clean', 'jacocoTestReport')
+        result = runVer(GRADLE_VERSION, 'clean', 'jacocoTestReport', '--warning-mode', 'all')
 
         then: "task done"
         result.task(":jacocoTestReport").outcome == TaskOutcome.SUCCESS
-        result.task(":jacocoMerge").outcome == TaskOutcome.SUCCESS
 
         then: "coverage aggregated"
-        file('build/reports/jacoco/test/jacocoTestReport.xml').exists()
-        file('build/jacoco/test.exec').exists()
+        def cov = file('build/reports/jacoco/test/jacocoTestReport.xml')
+        cov.exists()
+        cov.length() > 0
 
         when: "run dependencies task"
-        result = runVer(GRADLE_VERSION, 'htmlDependencyReport')
+        result = runVer(GRADLE_VERSION, 'htmlDependencyReport', '--warning-mode', 'all')
 
         then: "task done"
         result.task(":htmlDependencyReport").outcome == TaskOutcome.SUCCESS